Changes the online storage to match http
The blockchain-http repo already had a workaround in place to return a hotpot status as online if the hotspots had any activity on chain.

This worked but would then not reflect in generated extracts or in city counts.

This PR updates the gateway_status.online column based on the same logic.. i.e.: A hotspot is _online_ if it has _any_ on chain activity in the last 36 hours of blocks (2160 blocks).

A side effect of this PR is that ARPs are no longer sent since we’re no longer interested in using the hotspot gossiped height as an indication of how far along it is. This is not a big deal _except_ that listen_addresses are no longer updated as frequently which will no longer matter with light hotspots around the corner