uddate last_block for gateway inventory
There were cases where the ledger hook got triggered for things that are unreleated to what the database considers material changes to gateway information.

This now validates for unhandled gateways (i.e. not already handled in the block being processed) that relevant gateway of block data  is actually changed. This is only the reward_scale in the current chain.