Adding automation for building and publishing public Docker images.
This resolves issue #270 by creating a new Docker image and publishing it to the GitHub package repo. This may also resolve #265 by providing public containers when a release happens so users don't have to build their own.

This works off git tags and anytime an image is tagged in the format of x.y.z the workflow will be kicked off. I chose to use the GH package repo as to not interfere with any currently in-place systems with the quay.io hosting.